\name{nci}
\alias{nci}
\title{NCI pathways}
\description{
  A list with an entry for each NCI pathway.
 
  Schaefer CF, Anthony K, Krupa S, Buchoff J, Day M, Hannay T, Buetow KH. PID:
the Pathway Interaction Database. Nucleic Acids Res. 2009 Jan;37(Database
issue):D674-9. Epub 2008 Oct 2.

  \url{http://pid.nci.nih.gov/}
}
