\name{Pathways-class}
\Rdversion{1.1}
\docType{class}
\alias{Pathways-class}

\title{Class \code{"Pathways"}}
\description{
  A virtual class acting as a common parent to all other classes
  representing pathway databases.
}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.}
\section{Methods}{
  No methods defined with class "Pathways" in the signature.
}
\author{
  Gabriele Sales
}

\seealso{
  \code{\linkS4class{PathwayList}}
}
\keyword{classes}
