\name{reactome}
\alias{reactome}
\title{Reactome pathways}
\description{
  A list with an entry for each Reactome pathway.
 
  Matthews L, Gopinath G, Gillespie M, Caudy M, Croft D, de Bono B, Garapati P, 
Hemish J, Hermjakob H, Jassal B, Kanapin A, Lewis S, Mahajan S, May B, Schmidt E,
Vastrik I, Wu G, Birney E, Stein L, D'Eustachio P. Reactome knowledgebase of
human biological pathways and processes. Nucleic Acids Res. 2009 Jan;37(Database 
issue):D619-22. Epub 2008 Nov 3.

  \url{http://www.reactome.org}
}
